move loading spinner to separate component
