+++ components/PostChooser/editor-partials/loading-overlay/editor.scss
@@ -1,0 +1,54 @@
+// This overlay is used to indicate that results are being loaded or searched.
+.bu-components-post-chooser-results-overlay {
+	opacity: 0;
+	pointer-events: none;
+	position: absolute;
+	transition: opacity 300ms ease-in-out;
+	z-index: 2;
+	left: -12px;
+	right: -12px;
+	bottom: -12px;
+	top: -12px;
+	background: rgba(0, 0, 0, 0.125);
+	backdrop-filter: blur(1px);
+
+	&[data-overlayvisible="true"] {
+		opacity: 1;
+	}
+	&[data-overlayvisible="false"] {
+		opacity: 0;
+	}
+
+}
+
+
+
+// Container for the Loading Spinner in the Results component.
+// Sits on top of the results list.
+// This spinner is used to indicate that results are being loaded or searched.
+.bu-components-post-chooser-results-spinner {
+	opacity: 0;
+	pointer-events: none;
+	position: absolute;
+	transition: opacity 300ms ease-in-out;
+	z-index: 3;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	top: 0;
+
+	&[data-spinnervisible="true"] {
+		opacity: 1;
+	}
+	&[data-spinnervisible="false"] {
+		opacity: 0;
+	}
+
+	.bu-components-loading-spinner {
+		left: 50%;
+		transform: translate(-50%, -50%);
+		position: absolute;
+		top: 50%;
+		z-index: 4;
+	}
+}
